마크 12.0 (플레이어 체력바 구현)
각종 NULL 버그 고침, 플레이어 스텟하고 적 스텟 파씽해서 하면 좋을 듯 나중에 회의 때나 다같이 있을 때 물어봐야 될 듯, 박현욱 바보

+++ HoneyComb/Assets/02.Scripts/weapon.cs
@@ -8,7 +8,7 @@ public class weapon : MonoBehaviour
     //공격 구분짓기 위함
     public int id; //무기 or 공격 아이디
     public int prefabId; //프리펩 아이디
-    public float damage; //데미지
+    public int damage; //데미지
     public int count; //관통 되는 수
     public float speed; // 공속
